Switched js and css to modular system
Organized into component folders instead of using top-level tag imports in report.html
